Library: `lib::Several` complete and tested (see #473)
As a replacement for the `RefArray` a new generic container
has been implemented and tested, in interplay with `AllocationCluster`
 * the front-end container `lib::Several<I>` exposes only a reference
   to the ''interface type'' `I`, while hiding any storage details
 * data can only be populated through the `lib::SeveralBuilder`
 * a lot of flexibility is allowed for the actual element data types
 * element storage is maintained in a storage extent, managed through
   a custom allocator (defaulting to `std::allocator` ⟹ heap storage)

+++ src/lib/allocation-cluster.hpp
@@ -366,7 +366,7 @@ namespace lib {
               {
                 if (not bucket) return false;
                 size_t currSize = bucket->getAllocSize();
-                size_t delta = request - bucket->buffSiz;
+                long delta = long(request) - long(bucket->buffSiz);
                 return this->mother_->canAdjust (bucket,currSize, currSize+delta);
               }
